rename plotting scripts
